Cleanup: move form layout components to common/LayoutComponents
