+++ ui/src/components/common/LayoutComponents/index.ts
@@ -1,5 +1,8 @@
 export * from './Column';
 export * from './Container';
+export * from './FormColumn';
+export * from './FormContainer';
+export * from './FormRow';
 export * from './HorizontalSeparator';
 export * from './Row';
 export * from './Visible';
