+++ ui/src/components/forms/common/ChangeValidityForm.tsx
@@ -2,9 +2,8 @@ import { FieldValues, Path } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { TranslationKey } from '../../../i18n';
 import { Priority } from '../../../types/enums';
-import { Row } from '../../common/LayoutComponents';
+import { FormRow, Row } from '../../common/LayoutComponents';
 import { FormState } from './ChangeValidityForm.schema';
-import { FormRow } from './FormRow';
 import { InputField } from './InputField';
 import { PriorityForm } from './PriorityForm';
 import { ValidityPeriodForm } from './ValidityPeriodForm';
