+++ ui/src/components/forms/common/index.ts
@@ -6,8 +6,6 @@ export * from './dirtyFields';
 export * from './DateInputField';
 export * from './FormActionButtons';
 export * from './DateRange';
-export * from './FormColumn';
-export * from './FormRow';
 export * from './InputElement';
 export * from './InputField';
 export * from './InputLabel';
