+++ ui/src/components/forms/route/TerminusNameInputs.tsx
@@ -1,7 +1,8 @@
 import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
-import { FormColumn, FormRow, InputField } from '../common';
+import { FormColumn, FormRow } from '../../common/LayoutComponents';
+import { InputField } from '../common';
 import { RouteFormState } from './RoutePropertiesForm.types';
 
 const testIds = {
