+++ ui/src/components/forms/stop/components/Location.tsx
@@ -1,5 +1,6 @@
 import { FC } from 'react';
-import { FormColumn, FormRow, InputField } from '../../common';
+import { FormColumn, FormRow } from '../../../common/LayoutComponents';
+import { InputField } from '../../common';
 import { StopFormState } from '../types';
 import { TimingPlace } from './TimingPlace';
