+++ ui/src/components/map/stop-areas/StopAreaNames.tsx
@@ -4,7 +4,8 @@ import { useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { accordionClassNames } from '../../common';
 import { ExpandButton } from '../../common/Buttons';
-import { FormRow, InputField } from '../../forms/common';
+import { FormRow } from '../../common/LayoutComponents';
+import { InputField } from '../../forms/common';
 import { StopAreaFormState } from '../../forms/stop-area';
 import { StopAreaEnglishNames } from './StopAreaEnglishNames';
