+++ ui/src/components/routes-and-lines/line-details/ActionsRow.tsx
@@ -1,9 +1,8 @@
 import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { SimpleButton } from '../../common/Buttons';
-import { Column, Container } from '../../common/LayoutComponents';
+import { Column, Container, FormRow } from '../../common/LayoutComponents';
 import { ObservationDateControl } from '../../common/ObservationDateControl';
-import { FormRow } from '../../forms/common';
 import { useRoutesAndLinesDraftReturnToQueryParam } from '../line-drafts/useRoutesAndLinesDraftReturnToQueryParam';
 import { useGetLineDetails } from './useGetLineDetails';
