+++ ui/src/components/stop-registry/stops/stop-details/hybrid-stop/mirrored-quay-form/MirroredQuayBasicDetailsForm.tsx
@@ -7,10 +7,9 @@ import { mapStopPlaceStateToUiName } from '../../../../../../i18n/uiNameMappings
 import { StopWithDetails } from '../../../../../../types';
 import { StopPlaceState } from '../../../../../../types/stop-registry';
 import { EnumDropdown } from '../../../../../common/Dropdowns';
+import { FormColumn, FormRow } from '../../../../../common/LayoutComponents';
 import {
   FormActionButtons,
-  FormColumn,
-  FormRow,
   InputElement,
   InputField,
   ReasonForChangeForm,
