+++ ui/src/components/stop-registry/stops/stop-details/maintenance/OrganisationDetailsForm.tsx
@@ -5,8 +5,8 @@ import { FormProvider, useForm } from 'react-hook-form';
 import { z } from 'zod';
 import { submitFormByRef } from '../../../../../utils';
 import { SimpleButton } from '../../../../common/Buttons';
-import { Row, Visible } from '../../../../common/LayoutComponents';
-import { FormRow, InputField, requiredString } from '../../../../forms/common';
+import { FormRow, Row, Visible } from '../../../../common/LayoutComponents';
+import { InputField, requiredString } from '../../../../forms/common';
 import { useDirtyFormBlockNavigation } from '../../../../forms/common/NavigationBlocker';
 
 const testIds = {
