+++ ui/src/components/stop-registry/stops/stop-details/stop-version/StopVersionForm.tsx
@@ -3,9 +3,8 @@ import { useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
 import { SimpleButton } from '../../../../common/Buttons';
-import { Row } from '../../../../common/LayoutComponents';
+import { FormRow, Row } from '../../../../common/LayoutComponents';
 import {
-  FormRow,
   PriorityForm,
   ReasonForChangeForm,
   ValidityPeriodForm,
