+++ ui/src/components/stop-registry/terminals/components/basic-details/basic-details-form/TerminalDetailsEdit.tsx
@@ -18,13 +18,12 @@ import {
   showSuccessToast,
 } from '../../../../../../utils';
 import { useLoader } from '../../../../../common/hooks/useLoader';
-import { Column } from '../../../../../common/LayoutComponents';
 import {
-  FormActionButtons,
+  Column,
   FormColumn,
   FormRow,
-  InputField,
-} from '../../../../../forms/common';
+} from '../../../../../common/LayoutComponents';
+import { FormActionButtons, InputField } from '../../../../../forms/common';
 import { useDirtyFormBlockNavigation } from '../../../../../forms/common/NavigationBlocker';
 import { AlternativeNamesEdit } from '../../../../components/AlternativeNames/AlternativeNamesEdit';
 import { SelectedStop } from '../../../../components/SelectMemberStops/common/schema';
