+++ ui/src/components/stop-registry/terminals/components/owner-details/OwnerDetailsEdit.tsx
@@ -6,12 +6,8 @@ import { Operation } from '../../../../../redux';
 import { EnrichedParentStopPlace } from '../../../../../types';
 import { showSuccessToast } from '../../../../../utils';
 import { useLoader } from '../../../../common/hooks/useLoader';
-import { Column } from '../../../../common/LayoutComponents';
-import {
-  FormActionButtons,
-  FormRow,
-  InputField,
-} from '../../../../forms/common';
+import { Column, FormRow } from '../../../../common/LayoutComponents';
+import { FormActionButtons, InputField } from '../../../../forms/common';
 import { useDirtyFormBlockNavigation } from '../../../../forms/common/NavigationBlocker';
 import { OwnerOrganizationFields } from './OwnerOrganisationFields';
 import {
