+++ ui/src/components/timetables/common/ChangeTimetablesValidityForm.tsx
@@ -5,13 +5,13 @@ import { useTranslation } from 'react-i18next';
 import { z } from 'zod';
 import { submitFormByRef } from '../../../utils';
 import { SimpleButton } from '../../common/Buttons';
-import { Row, Visible } from '../../common/LayoutComponents';
 import {
   FormColumn,
   FormRow,
-  InputField,
-  requiredString,
-} from '../../forms/common';
+  Row,
+  Visible,
+} from '../../common/LayoutComponents';
+import { InputField, requiredString } from '../../forms/common';
 import { useDirtyFormBlockNavigation } from '../../forms/common/NavigationBlocker';
 import { AffectedRouteLabels } from './AffectedRouteLabels';
