+++ ui/src/components/timetables/import/ImportTimetablesPage.tsx
@@ -15,8 +15,7 @@ import { ConfirmationDialog } from '../../../uiComponents';
 import { showDangerToastWithError, showSuccessToast } from '../../../utils';
 import { PageTitle } from '../../common';
 import { CloseIconButton, SimpleButton } from '../../common/Buttons';
-import { Container, Row } from '../../common/LayoutComponents';
-import { FormRow } from '../../forms/common';
+import { Container, FormRow, Row } from '../../common/LayoutComponents';
 import { ConfirmTimetablesImportModal } from './ConfirmTimetablesImportModal';
 import { FileImportDragAndDrop } from './FileImportDragAndDrop';
 import { useTimetablesImport } from './hooks';
