+++ ui/src/components/timetables/substitute-day-settings/OccasionalSubstitutePeriod/OccasionalSubstitutePeriodRow.tsx
@@ -3,8 +3,13 @@ import { FieldArrayWithId, useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { MdDelete, MdUndo } from 'react-icons/md';
 import { SimpleButton } from '../../../common/Buttons';
-import { Column, Visible } from '../../../common/LayoutComponents';
-import { FormColumn, FormRow, InputField } from '../../../forms/common';
+import {
+  Column,
+  FormColumn,
+  FormRow,
+  Visible,
+} from '../../../common/LayoutComponents';
+import { InputField } from '../../../forms/common';
 import { LineTypeMultiSelectDropdown } from '../../../forms/timetables/LineTypeMultiSelectDropdown';
 import { SubstituteDayOfWeekDropdown } from '../../../forms/timetables/SubstituteDayOfWeekDropdown';
 import { FormState, PeriodType } from './OccasionalSubstitutePeriodForm.types';
